Former-commit-id: 0697e90bf3caee997475f74dab0ee4d2332e16bf

+++ user/psava/Mcic2d.c
@@ -37,7 +37,7 @@ int main(int argc, char* argv[])
     sf_axis az,ax,at,aa; /* cube axes */
     int     iz,ix,it;
     int     nz,nx,nt;
-    unsigned int iseek;
+    off_t iseek;
 
     float **us=NULL,**ur=NULL,**ii=NULL;
